+++ FourSquareApp/View Model/SearchVm.swift
@@ -121,7 +121,7 @@ class SearchViewModel {
                                     if let data07 = i["priceRange"] as? Int{
                                         priceRangeIs = String(data07)
                                     }
-                                    if let data08 = i["rating"] as? Float{
+                                    if let data08 = i["rating"] as? Double{
                                         let number = Float(data08)
                                         ratingIs = String(format: "%.1f", number)
                                     }
@@ -209,7 +209,7 @@ class SearchViewModel {
                                     if let data07 = i["priceRange"] as? Int{
                                         priceRangeIs = String(data07)
                                     }
-                                    if let data08 = i["rating"] as? Float{
+                                    if let data08 = i["rating"] as? Double{
                                         let number = Float(data08)
                                         ratingIs = String(format: "%.1f", number)
                                     }
